refactor helper methods to ContentDirectoryCommandBase
