+++ cli/src/commands/content/deleteChannel.ts
@@ -1,7 +1,6 @@
 import ContentDirectoryCommandBase from '../../base/ContentDirectoryCommandBase'
 import { flags } from '@oclif/command'
 import chalk from 'chalk'
-import { createType } from '@joystream/types'
 import ExitCodes from '../../ExitCodes'
 import { formatBalance } from '@polkadot/util'
 import BN from 'bn.js'
@@ -21,36 +20,6 @@ export default class DeleteChannelCommand extends ContentDirectoryCommandBase {
       description: 'Force-remove all associated channel data objects',
     }),
     ...ContentDirectoryCommandBase.flags,
-  }
-
-  async getDataObjectsInfoFromQueryNode(channelId: number): Promise<[string, BN][]> {
-    const dataObjects = await this.getQNApi().dataObjectsByBagId(`dynamic:channel:${channelId}`)
-
-    if (dataObjects.length) {
-      this.log('Following data objects are still associated with the channel:')
-      dataObjects.forEach((o) => {
-        let parentStr = ''
-        if ('video' in o.type && o.type.video) {
-          parentStr = ` (video: ${o.type.video.id})`
-        }
-        this.log(`- ${o.id} - ${o.type.__typename}${parentStr}`)
-      })
-    }
-
-    return dataObjects.map((o) => [o.id, new BN(o.stateBloatBond)])
-  }
-
-  async getDataObjectsInfoFromChain(channelId: number): Promise<[string, BN][]> {
-    const dataObjects = await this.getApi().dataObjectsInBag(
-      createType('PalletStorageBagIdType', { Dynamic: { Channel: channelId } })
-    )
-
-    if (dataObjects.length) {
-      const dataObjectIds = dataObjects.map(([id]) => id.toString())
-      this.log(`Following data objects are still associated with the channel: ${dataObjectIds.join(', ')}`)
-    }
-
-    return dataObjects.map(([id, o]) => [id.toString(), o.stateBloatBond])
   }
 
   async run(): Promise<void> {
